+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.1.0-alpha.6",
-  "contentSha256_12": "be7973b707f9",
+  "pkgVersion": "1.1.0-alpha.7",
+  "contentSha256_12": "8c7ca44de83c",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.1.0-alpha.6";
-export const contentSha256_12 = "be7973b707f9";
+export const pkgVersion = "1.1.0-alpha.7";
+export const contentSha256_12 = "8c7ca44de83c";
 export const file = "nodalia-cards.bundle.js";